+++ src/server/ssr.ts
@@ -1,11 +1,4 @@
-import {
-  html,
-  unsafeHTML,
-  render,
-  collectResult,
-  DOMParser,
-  Element,
-} from "../deps.ts";
+import { html, unsafeHTML, render, collectResult, DOMParser } from "../deps.ts";
 import type { RouterContext } from "../deps.ts";
 // @ts-ignore lit is a npm package and Deno doesn't resolve the exported members
 import type { LitElement, TemplateResult } from "lit";
@@ -19,11 +12,17 @@ import type { ComponentContext } from "./router.ts";
 
 import "../runtime/is-land.ts"; // should use limette?
 
+import { installWindowOnGlobal } from "../deps.ts";
+installWindowOnGlobal();
+// Set window object, because the shim doesn't do it
+// @ts-ignore some components use the `window` reference for registration process
+globalThis.window = globalThis;
+
 type Params = {
   [key: string]: string;
 };
 
-export type AppTemplateOptions = {
+export type AppRootOptions = {
   css: DirectiveResult<UnsafeHTMLDirective> | string;
   js: string[] | TemplateResult[] | DirectiveResult<UnsafeHTMLDirective>[];
   component: DirectiveResult<UnsafeHTMLDirective>;
@@ -36,11 +35,11 @@ export declare class ComponentCtxMixinInterface {
 
 export declare class AppTemplateInterface {
   prototype: {
-    render(app: AppTemplateOptions): TemplateResult;
+    render(app: AppRootOptions): TemplateResult;
   };
 }
 
-function registerComponent(
+function registerRouteComponent(
   ComponentClass: CustomElementConstructor,
   tagName: string
 ) {
@@ -125,79 +124,6 @@ function moveLmtHeadElements(htmlString: string) {
   return doc.documentElement.outerHTML;
 }
 
-function moveLmtHeadElementsFirstLevel(htmlString: string) {
-  const parser = new DOMParser();
-  const doc = parser.parseFromString(
-    htmlString,
-    "text/html"
-  ) as unknown as Document;
-
-  const uniqueElements = new Map(); // Track the last unique element by key
-  let lastTitleElement = null; // Track the last <title> element found
-
-  function processLmtHeadElements(root: Document | DocumentFragment) {
-    // Find all <lmt-head> elements within the given root
-    const lmtHeadElements = Array.from(root.querySelectorAll("lmt-head"));
-
-    lmtHeadElements.forEach((wrapper) => {
-      Array.from(wrapper.children).forEach((child) => {
-        if (child.tagName === "TITLE") {
-          // Keep only the last <title> element found
-          lastTitleElement = child;
-        } else {
-          const key = child.getAttribute("key");
-
-          if (key) {
-            // Only keep the latest element with each key
-            uniqueElements.set(key, child);
-          } else {
-            // Treat children without a key as unique
-            uniqueElements.set(Symbol(), child);
-          }
-        }
-      });
-
-      // Remove the <lmt-head> wrapper from the DOM
-      wrapper.remove();
-    });
-  }
-
-  // Process <lmt-head> elements in the main DOM
-  processLmtHeadElements(doc);
-
-  // Process <lmt-head> elements within each declarative shadow DOM template
-  const templates = doc.querySelectorAll(
-    "template[shadowroot]"
-  ) as unknown as HTMLTemplateElement[];
-  templates.forEach((template) => {
-    processLmtHeadElements(template.content); // Directly use template.content without re-parsing
-  });
-
-  // Remove redundant elements in <head> that match keys in uniqueElements
-  Array.from(doc.head.querySelectorAll("[key]")).forEach((headElement) => {
-    const key = headElement.getAttribute("key");
-    if (key && uniqueElements.has(key)) {
-      headElement.remove();
-    }
-  });
-
-  // Remove any existing <title> elements in <head> if we have a new one
-  if (lastTitleElement) {
-    Array.from(doc.head.querySelectorAll("title")).forEach((title) =>
-      title.remove()
-    );
-    // Append the last <title> element found to <head>
-    doc.head.appendChild(lastTitleElement);
-  }
-
-  // Append the deduplicated elements to <head>
-  uniqueElements.forEach((element) => {
-    doc.head.appendChild(element);
-  });
-
-  return doc.documentElement.outerHTML;
-}
-
 const ComponentCtxMixin = (base: typeof LitElement) => {
   class ComponentCtxClass extends base {
     __ctx: ComponentContext = { params: {}, data: undefined };
@@ -211,7 +137,7 @@ const ComponentCtxMixin = (base: typeof LitElement) => {
 };
 
 export function bootstrapContent(
-  AppTemplate: AppTemplateInterface,
+  AppRoot: AppTemplateInterface,
   route: BuildRoute,
   componentContext: ComponentContext
 ) {
@@ -220,7 +146,7 @@ export function bootstrapContent(
   const ComponentClass = ComponentCtxMixin(
     componentModule?.default as typeof LitElement
   );
-  const component = registerComponent(
+  const component = registerRouteComponent(
     ComponentClass as unknown as CustomElementConstructor,
     route.tagName
   );
@@ -229,7 +155,7 @@ export function bootstrapContent(
     componentContext
   )}</script>`;
 
-  const appTemplateOptions: AppTemplateOptions = {
+  const appTemplateOptions: AppRootOptions = {
     css: route.cssAssetPath
       ? unsafeHTML(`<link rel="stylesheet" href="${route.cssAssetPath}" />`)
       : ``,
@@ -242,11 +168,11 @@ export function bootstrapContent(
     component: unsafeHTML(component),
   };
 
-  return AppTemplate.prototype.render(appTemplateOptions);
+  return AppRoot.prototype.render(appTemplateOptions);
 }
 
 export async function renderContent(
-  AppTemplate: AppTemplateInterface,
+  AppRoot: AppTemplateInterface,
   route: BuildRoute,
   routerContext: RouterContext<typeof route.path>,
   data?: ComponentContext["data"]
@@ -255,7 +181,7 @@ export async function renderContent(
 
   const result = render(
     await bootstrapContent(
-      AppTemplate as AppTemplateInterface,
+      AppRoot as AppTemplateInterface,
       route,
       componentContext
     ),